feat: [FC-86] implement multi-value faceted search  (#213)
This is implemented in both the Meilisearch and Elasticsearch backends.

Course discovery needs to support filtering courses by multiple options
within a single facet category. This is necessary to support the MFE
conversion of the Catalog.

Prior Behavior:
Using the "language" facet as an example. After selecting "English" in
the Language facet category, the language section in the filters area
only displays the selected value - "English".

New Behavior:
After selecting "English" in the Language facet category, the language
section in the filters area displays all language values ("English",
"French", "Ukrainian", etc.) as long as they are available considering
the rest of the currently selected filters (organization, mode).

This change started as Meilisearch-specific, but this led to an issue
with running the tests. The api layer abstracts away the selection of
the search engine, which means the new logic automatically affects the
Elasticsearch-specific parts and some tests start to fail.
Consequently, it was decided to make this commit cover both Meilisearch
and Elasticsearch.

---------

Co-authored-by: Іван Нєдєльніцев <ivan.niedielnitsev@raccoongang.com>